+++ Package.swift
@@ -8,9 +8,9 @@ let package = Package(
       dependencies: [.Target(name: "SourceKittenFramework")]),
   ],
   dependencies: [
-    .Package(url: "https://github.com/Carthage/Commandant.git", versions: Version(0, 11, 3)..<Version(0, 11, .max)),
+    .Package(url: "https://github.com/Carthage/Commandant.git", versions: Version(0, 12, 0)..<Version(0, 12, .max)),
     .Package(url: "https://github.com/drmohundro/SWXMLHash.git", Version(3, 0, 4)),
-    .Package(url: "https://github.com/jpsim/Yams.git", Version(0, 2, 0)),
+    .Package(url: "https://github.com/jpsim/Yams.git", Version(0, 3, 0)),
     .Package(url: "https://github.com/norio-nomura/Clang_C.git", majorVersion: 1),
     .Package(url: "https://github.com/norio-nomura/SourceKit.git", majorVersion: 1),
   ],